Refactor serialize module
