+++ utilities/src/serialize/flags.rs
@@ -14,15 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-pub trait Flags: Default + Clone + Copy + Sized {
-    fn u8_bitmask(&self) -> u8;
-
-    fn from_u8(value: u8) -> Self;
-
-    fn from_u8_remove_flags(value: &mut u8) -> Self;
-
-    fn len() -> usize;
-}
+use crate::serialize::Flags;
 
 /// Flags to be encoded into the serialization.
 #[derive(Default, Clone, Copy)]
